refactor: generateLottoNumbersSet 분리
